requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.4
Django==4.1
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.21.4
factory-boy==3.2.1
Faker==14.1.0
flake8==5.0.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
openapi-codec==1.3.2
packaging==21.3
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
